requirements.txt
Flask
psycopg2-binary
Flask-Session
werkzeug
